flaskbook/requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
pymongo==2.9
mongoengine==0.8
flask-mongoengine==0.7.4
py-bcrypt==0.4